+++ test/snippets/snippets.cpp
@@ -129,648 +129,6 @@ void run_coro(boost::asio::any_io_executor ex, std::function<boost::asio::awaita
 }
 #endif
 
-static std::string get_name() { return "John"; }
-
-#ifndef BOOST_NO_CXX17_HDR_OPTIONAL
-//[sql_formatting_incremental_fn
-// Compose an update query that sets first_name, last_name, or both
-std::string compose_update_query(
-    boost::mysql::format_options opts,
-    std::int64_t employee_id,
-    std::optional<std::string> new_first_name,
-    std::optional<std::string> new_last_name
-)
-{
-    // There should be at least one update
-    assert(new_first_name || new_last_name);
-
-    // format_context will accumulate the query as we compose it
-    boost::mysql::format_context ctx(opts);
-
-    // append_raw adds raw SQL to the generated query, without quoting or escaping.
-    // You can only pass strings known at compile-time to append_raw,
-    // unless you use the runtime function.
-    ctx.append_raw("UPDATE employee SET ");
-
-    if (new_first_name)
-    {
-        // format_sql_to expands a format string and appends the result
-        // to a format context. This way, we can build our query in small pieces
-        // Add the first_name update clause
-        boost::mysql::format_sql_to(ctx, "first_name = {}", *new_first_name);
-    }
-    if (new_last_name)
-    {
-        if (new_first_name)
-            ctx.append_raw(", ");
-
-        // Add the last_name update clause
-        boost::mysql::format_sql_to(ctx, "last_name = {}", *new_last_name);
-    }
-
-    // Add the where clause
-    boost::mysql::format_sql_to(ctx, " WHERE id = {}", employee_id);
-
-    // Retrieve the generated query string
-    return std::move(ctx).get().value();
-}
-//]
-#endif
-
-//[sql_formatting_formatter_specialization
-// We want to add formatting support for employee_t
-struct employee_t
-{
-    std::string first_name;
-    std::string last_name;
-    std::string company_id;
-};
-
-namespace boost {
-namespace mysql {
-
-template <>
-struct formatter<employee_t>
-{
-    // formatter<T> should define, at least, a function with signature:
-    //    static void format(const T&, format_context_base&)
-    // This function must use format_sql_to, format_context_base::append_raw
-    // or format_context_base::append_value to format the passed value.
-    // We will make this suitable for INSERT statements
-    static void format(const employee_t& emp, format_context_base& ctx)
-    {
-        format_sql_to(ctx, "{}, {}, {}", emp.first_name, emp.last_name, emp.company_id);
-    }
-};
-
-}  // namespace mysql
-}  // namespace boost
-//]
-
-#ifndef BOOST_NO_CXX17_HDR_OPTIONAL
-//[sql_formatting_unit_test
-// For reference, the function under test
-std::string compose_update_query(
-    boost::mysql::format_options opts,
-    std::int64_t employee_id,
-    std::optional<std::string> new_first_name,
-    std::optional<std::string> new_last_name
-);
-
-// Your test body
-void test_compose_update_query()
-{
-    // You can safely use these format_options for testing,
-    // since they are the most common ones.
-    boost::mysql::format_options opts{boost::mysql::utf8mb4_charset, true};
-
-    // Test for the different cases
-    ASSERT(
-        compose_update_query(opts, 42, "Bob", {}) == "UPDATE employee SET first_name = 'Bob' WHERE id = 42"
-    );
-    ASSERT(
-        compose_update_query(opts, 42, {}, "Alice") == "UPDATE employee SET last_name = 'Alice' WHERE id = 42"
-    );
-    ASSERT(
-        compose_update_query(opts, 0, "Bob", "Alice") ==
-        "UPDATE employee SET first_name = 'Bob', last_name = 'Alice' WHERE id = 0"
-    );
-}
-//]
-#endif
-
-void section_sql_formatting()
-{
-    boost::mysql::connect_params params;
-    params.server_address.emplace_host_and_port(server_hostname);
-    params.username = mysql_username;
-    params.password = mysql_password;
-    params.database = "boost_mysql_examples";
-    params.multi_queries = true;
-    params.ssl = boost::mysql::ssl_mode::disable;
-
-    boost::asio::io_context ioc;
-    boost::mysql::any_connection conn(ioc);
-    boost::mysql::results r;
-
-    conn.connect(params);
-
-    {
-        //[sql_formatting_simple
-        std::string employee_name = get_name();  // employee_name is an untrusted string
-
-        // Compose the SQL query in the client
-        std::string query = boost::mysql::format_sql(
-            conn.format_opts().value(),
-            "SELECT id, salary FROM employee WHERE last_name = {}",
-            employee_name
-        );
-
-        // If employee_name is "John", query now contains:
-        // "SELECT id, salary FROM employee WHERE last_name = 'John'"
-        // If employee_name contains quotes, they will be escaped as required
-
-        // Execute the generated query as usual
-        results result;
-        conn.execute(query, result);
-        //]
-
-        ASSERT(query == "SELECT id, salary FROM employee WHERE last_name = 'John'");
-    }
-    {
-        //[sql_formatting_other_scalars
-        std::string query = boost::mysql::format_sql(
-            conn.format_opts().value(),
-            "SELECT id FROM employee WHERE salary > {}",
-            42000
-        );
-
-        ASSERT(query == "SELECT id FROM employee WHERE salary > 42000");
-        //]
-
-        conn.execute(query, r);
-    }
-#ifndef BOOST_NO_CXX17_HDR_OPTIONAL
-    {
-        //[sql_formatting_optionals
-        std::optional<std::int64_t> salary;  // get salary from a possibly untrusted source
-
-        std::string query = boost::mysql::format_sql(
-            conn.format_opts().value(),
-            "UPDATE employee SET salary = {} WHERE id = {}",
-            salary,
-            1
-        );
-
-        // Depending on whether salary has a value or not, generates:
-        // UPDATE employee SET salary = 42000 WHERE id = 1
-        // UPDATE employee SET salary = NULL WHERE id = 1
-        //]
-
-        ASSERT(query == "UPDATE employee SET salary = NULL WHERE id = 1");
-        conn.execute(query, r);
-    }
-#endif
-    {
-        //[sql_formatting_manual_indices
-        // Recall that you need to set connect_params::multi_queries to true when connecting
-        // before running semicolon-separated queries.
-        std::string query = boost::mysql::format_sql(
-            conn.format_opts().value(),
-            "UPDATE employee SET first_name = {1} WHERE id = {0}; SELECT * FROM employee WHERE id = {0}",
-            42,
-            "John"
-        );
-
-        ASSERT(
-            query ==
-            "UPDATE employee SET first_name = 'John' WHERE id = 42; SELECT * FROM employee WHERE id = 42"
-        );
-        //]
-
-        conn.execute(query, r);
-    }
-    {
-        // clang-format off
-        //[sql_formatting_named_args
-        std::string query = boost::mysql::format_sql(
-            conn.format_opts().value(),
-            "UPDATE employee SET first_name = {name} WHERE id = {id}; SELECT * FROM employee WHERE id = {id}",
-            {
-                {"id",   42    },
-                {"name", "John"}
-            }
-        );
-        //<-
-        // clang-format on
-        //->
-
-        ASSERT(
-            query ==
-            "UPDATE employee SET first_name = 'John' WHERE id = 42; SELECT * FROM employee WHERE id = 42"
-        );
-        //]
-
-        conn.execute(query, r);
-    }
-    {
-        //[sql_formatting_identifiers
-        std::string query = boost::mysql::format_sql(
-            conn.format_opts().value(),
-            "SELECT id, last_name FROM employee ORDER BY {} DESC",
-            boost::mysql::identifier("company_id")
-        );
-
-        ASSERT(query == "SELECT id, last_name FROM employee ORDER BY `company_id` DESC");
-        //]
-
-        conn.execute(query, r);
-    }
-    {
-        //[sql_formatting_qualified_identifiers
-        std::string query = boost::mysql::format_sql(
-            conn.format_opts().value(),
-            "SELECT salary, tax_id FROM employee "
-            "INNER JOIN company ON employee.company_id = company.id "
-            "ORDER BY {} DESC",
-            boost::mysql::identifier("company", "id")
-        );
-        // SELECT ... ORDER BY `company`.`id` DESC
-        //]
-
-        ASSERT(
-            query ==
-            "SELECT salary, tax_id FROM employee "
-            "INNER JOIN company ON employee.company_id = company.id "
-            "ORDER BY `company`.`id` DESC"
-        );
-        conn.execute(query, r);
-    }
-#ifndef BOOST_NO_CXX17_HDR_OPTIONAL
-    {
-        //[sql_formatting_incremental_use
-        std::string query = compose_update_query(conn.format_opts().value(), 42, "John", {});
-
-        ASSERT(query == "UPDATE employee SET first_name = 'John' WHERE id = 42");
-        //]
-
-        conn.execute(query, r);
-    }
-    {
-        test_compose_update_query();
-    }
-#endif
-    {
-        try
-        {
-            //[sql_formatting_invalid_encoding
-            // If the connection is using UTF-8 (the default), this will throw an error,
-            // because the string to be formatted contains invalid UTF8.
-            format_sql(conn.format_opts().value(), "SELECT {}", "bad\xff UTF-8");
-            //]
-
-            ASSERT(false);
-        }
-        catch (const boost::system::system_error& err)
-        {
-            ASSERT(err.code() == boost::mysql::client_errc::invalid_encoding);
-        }
-    }
-    {
-        using namespace boost::mysql;
-        using boost::optional;
-        const auto opts = conn.format_opts().value();
-
-        //[sql_formatting_reference_signed
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", 42) == "SELECT 42"
-            //<-
-        );
-        //->
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", -1) == "SELECT -1"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_unsigned
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", 42u) == "SELECT 42"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_bool
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", false) == "SELECT 0"
-            //<-
-        );
-        //->
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", true) == "SELECT 1"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_string
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", "Hello world") == "SELECT 'Hello world'"
-            //<-
-        );
-        //->
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", "Hello 'world'") == R"(SELECT 'Hello \'world\'')"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_blob
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", blob{0x00, 0x48, 0xff}) == R"(SELECT x'0048ff')"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_float
-        //<-
-        ASSERT(
-            //->
-            // Equivalent to format_sql(opts, "SELECT {}", static_cast<double>(4.2f))
-            // Note that MySQL uses doubles for all floating point literals
-            format_sql(opts, "SELECT {}", 4.2f) == "SELECT 4.199999809265137e+00"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_double
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", 4.2) == "SELECT 4.2e+00"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_date
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", date(2021, 1, 2)) == "SELECT '2021-01-02'"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_datetime
-        //<-
-        ASSERT(
-            // clang-format off
-            //->
-            format_sql(opts, "SELECT {}", datetime(2021, 1, 2, 23, 51, 14)) == "SELECT '2021-01-02 23:51:14.000000'"
-            //<-
-            // clang-format on
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_time
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", std::chrono::seconds(121)) == "SELECT '00:02:01.000000'"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_nullptr
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", nullptr) == "SELECT NULL"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_optional
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", optional<int>(42)) == "SELECT 42"
-            //<-
-        );
-        //->
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", optional<int>()) == "SELECT NULL"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_field
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", field(42)) == "SELECT 42"
-            //<-
-        );
-        //->
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", field("abc")) == "SELECT 'abc'"
-            //<-
-        );
-        //->
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {}", field()) == "SELECT NULL"
-            //<-
-        );
-        //->
-        //]
-
-        //[sql_formatting_reference_identifier
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {} FROM t", identifier("salary")) == "SELECT `salary` FROM t"
-            //<-
-        );
-        //->
-        //<-
-        ASSERT(
-            //->
-            format_sql(opts, "SELECT {} FROM t", identifier("sal`ary")) == "SELECT `sal``ary` FROM t"
-            //<-
-        );
-        //->
-        //<-
-        ASSERT(
-            // clang-format off
-            //->
-            format_sql(opts, "SELECT {} FROM t", identifier("mytable", "myfield")) == "SELECT `mytable`.`myfield` FROM t"
-            //<-
-            // clang-format on
-        );
-        //->
-        //<-
-        ASSERT(
-            // clang-format off
-            //->
-            format_sql(opts, "SELECT {} FROM t", identifier("mydb", "mytable", "myfield")) == "SELECT `mydb`.`mytable`.`myfield` FROM t"
-            //<-
-            // clang-format on
-        );
-        //->
-        //]
-    }
-
-    // Advanced section
-    {
-        //[sql_formatting_formatter_use
-        // We can now use employee as a built-in value
-        std::string query = boost::mysql::format_sql(
-            conn.format_opts().value(),
-            "INSERT INTO employee (first_name, last_name, company_id) VALUES ({}), ({})",
-            employee_t{"John", "Doe", "HGS"},
-            employee_t{"Rick", "Johnson", "AWC"}
-        );
-
-        ASSERT(
-            query ==
-            "INSERT INTO employee (first_name, last_name, company_id) VALUES "
-            "('John', 'Doe', 'HGS'), ('Rick', 'Johnson', 'AWC')"
-        );
-        //]
-
-        conn.execute(query, r);
-    }
-    {
-        const auto opts = conn.format_opts().value();
-
-        //[sql_formatting_auto_indexing
-        ASSERT(format_sql(opts, "SELECT {}, {}, {}", 42, "abc", nullptr) == "SELECT 42, 'abc', NULL");
-        //]
-    }
-    {
-        const auto opts = conn.format_opts().value();
-
-        //[sql_formatting_manual_auto_mix
-        try
-        {
-            // Mixing manual and auto indexing is illegal. This will throw an exception.
-            format_sql(opts, "SELECT {0}, {}", 42);
-            //<-
-            ASSERT(false);
-            //->
-        }
-        catch (const boost::system::system_error& err)
-        {
-            ASSERT(err.code() == boost::mysql::client_errc::format_string_manual_auto_mix);
-        }
-        //]
-    }
-    {
-        const auto opts = conn.format_opts().value();
-
-        //[sql_formatting_unused_args
-        // This is OK
-        std::string query = format_sql(opts, "SELECT {}", 42, "abc");
-        //]
-        ASSERT(query == "SELECT 42");
-    }
-    {
-        const auto opts = conn.format_opts().value();
-
-        //[sql_formatting_brace_literal
-        ASSERT(format_sql(opts, "SELECT 'Brace literals: {{ and }}'") == "SELECT 'Brace literals: { and }'");
-        //]
-    }
-    {
-        const auto opts = conn.format_opts().value();
-
-        //[sql_formatting_format_double_error
-        try
-        {
-            // We're trying to format a double infinity value, which is not
-            // supported by MySQL. This will throw an exception.
-            std::string formatted_query = format_sql(opts, "SELECT {}", HUGE_VAL);
-            //<-
-            ASSERT(false);
-            boost::ignore_unused(formatted_query);
-            //->
-        }
-        catch (const boost::system::system_error& err)
-        {
-            ASSERT(err.code() == boost::mysql::client_errc::unformattable_value);
-        }
-        //]
-    }
-    {
-        const auto opts = conn.format_opts().value();
-
-        //[sql_formatting_no_exceptions
-        // ctx contains an error code that tracks whether any error happened
-        boost::mysql::format_context ctx(opts);
-
-        // We're trying to format a infinity, which is an error. This
-        // will set the error state, but won't throw.
-        format_sql_to(ctx, "SELECT {}, {}", HUGE_VAL, 42);
-
-        // The error state gets checked at this point. Since it is set,
-        // res will contain an error.
-        boost::system::result<std::string> res = std::move(ctx).get();
-        ASSERT(!res.has_value());
-        ASSERT(res.has_error());
-        ASSERT(res.error() == boost::mysql::client_errc::unformattable_value);
-        // res.value() would throw an error, like format_sql would
-        //]
-    }
-#ifdef __cpp_lib_polymorphic_allocator
-    {
-        //[sql_formatting_custom_string
-        // Create a format context that uses std::pmr::string
-        boost::mysql::basic_format_context<std::pmr::string> ctx(conn.format_opts().value());
-
-        // Compose your query as usual
-        boost::mysql::format_sql_to(ctx, "SELECT * FROM employee WHERE id = {}", 42);
-
-        // Retrieve the query as usual
-        std::pmr::string query = std::move(ctx).get().value();
-        //]
-
-        ASSERT(query == "SELECT * FROM employee WHERE id = 42");
-        conn.execute(query, r);
-    }
-#endif
-    {
-        //[sql_formatting_memory_reuse
-        // we want to re-use memory held by storage
-        std::string storage;
-
-        // storage is moved into ctx by the constructor. If any memory
-        // had been allocated by the string, it will be re-used.
-        boost::mysql::format_context ctx(conn.format_opts().value(), std::move(storage));
-
-        // Use ctx as you normally would
-        boost::mysql::format_sql_to(ctx, "SELECT {}", 42);
-
-        // When calling get(), the string is moved out of the context
-        std::string query = std::move(ctx).get().value();
-        //]
-
-        ASSERT(query == "SELECT 42");
-    }
-}
-
 void main_impl()
 {
     //
